Fix: Cell color and table match
This is was a bit tricky to fix.
Previous implementation adds colors to a list of already
exisiting color and subcellular location mapping.
This could get heavy when a user starts adding and
changing a lot of subcellular locations and colors.
I instead updated the list with the new colors,
adding only new colors whena  new subcellular location
is added too.