Feat: Use `shiny.semantic` to improve app UI (#13)
* Feat: Adds app version using shiny.semantic

* Chore: Updates DESCRIPTION and NEWS

Dev version is just a suggestion.

* Chore: Adds dependencies for shiny app

* Feat: Dispalys cell type

Removes download handler in the meantime.

* Chore: Removes non-working css styles

* Chore: Removes unnecessary dependencies

* Chore: Removes more unnecessary styling

* Feat: Adds table of subcellular and color mapping

* Feat: Improves table display of colors

Only showing subcellular locations for now.
Table should display subcellular location names.

* Feat: Attempts to display better SL color mappings

* Feat: Improves subcellular color table

* Chore: Removes unused reactive object

* Feat: Uses original apps IDs

I think it is much easier to track changes with this.

I was able to make the table to work, but still needs more
styling.

Button added to clear the colors and selected subcellular
locations, but not yet working.

* Chore: Removes divider on vertical stack

* Feat: Places color selection on left of cell

I think this creates a much better flow.
Table should and color should now clear after pressing
the button.

* Feat: Specify min-width for table

I am trying to display block the swissbiopic to appear
bigger, but it is not working in browsers, i.e. user agent
stylesheet is being ignored..

* Chore: Removes unnecessary code

* Chore: Adds github icon for more intuitive link

* Chore: Adds vertical divider

* Chore: Makes more descriptive labels

* Fix: Cell color and table match

This is was a bit tricky to fix.
Previous implementation adds colors to a list of already
exisiting color and subcellular location mapping.
This could get heavy when a user starts adding and
changing a lot of subcellular locations and colors.
I instead updated the list with the new colors,
adding only new colors whena  new subcellular location
is added too.

* Fix: Removes non-existent object

* Chore: Uses correct subcellular spelling

* Chore: Removes non-working css style

* Feat: Makes inputs compact segments

This reduces their sizes and the size they occupy.

* Chore: Rearranges color input, table, and cell

I think this arrangment provides the most space for the
cell. It would still be best if we can force #swissbiopic
to display: block.