Feat: Gives cell a min-width
This is to make sure that the cell stays at a relatively
viewable size because it has a tendency to shrink until
it cannot be seen.
I also added a toast when user clicks the button to
download the code to inform them that the code is
workin already in the clipboard.
Reverted the download logic as I can't make shinyscreenshot
work with svg.