Chore: Adds CI (#19)
* Chore: Adds CI

* Chore: Adds rcmdcheck, lintr, and spelling to suggests

This is for the CI to work.

* Chore: Adds testthat as suggest dependency

This is for github workflow to work properly.

* Chore: Updates man

* Chore: Adds CI

* Chore: Adds rcmdcheck, lintr, and spelling to suggests

This is for the CI to work.

* Chore: Adds testthat as suggest dependency

This is for github workflow to work properly.

* Feat: Removes camelCase from lint

This is necessary to avoid lint on drawCell object name.

* Chore: Styles files to pass lint

* Chore: Suggests dev version; updates NEWS

* Chore: Updates drawCell's roxygen

* Chore: Update dependencies

* Chore: Adds wordlist for spell check

* Chore: Updates README

Disabled rendering of images for now because I am getting
an error with PhantomJS, which should not be used anyway.

* Chore: Fix spelling errors

Removed all reference to taxonomy_id in favor of the new
argument name ornganism_identifier.

* Chore: Returns suggests for CI

* Feat: Adds app version using shiny.semantic

* Chore: Updates DESCRIPTION and NEWS

Dev version is just a suggestion.

* Chore: Adds dependencies for shiny app

* Feat: Dispalys cell type

Removes download handler in the meantime.

* Chore: Removes non-working css styles

* Chore: Removes unnecessary dependencies

* Chore: Removes more unnecessary styling

* Feat: Adds table of subcellular and color mapping

* Feat: Improves table display of colors

Only showing subcellular locations for now.
Table should display subcellular location names.

* Feat: Attempts to display better SL color mappings

* Feat: Improves subcellular color table

* Chore: Removes unused reactive object

* Feat: Uses original apps IDs

I think it is much easier to track changes with this.

I was able to make the table to work, but still needs more
styling.

Button added to clear the colors and selected subcellular
locations, but not yet working.

* Chore: Removes divider on vertical stack

* Feat: Places color selection on left of cell

I think this creates a much better flow.
Table should and color should now clear after pressing
the button.

* Feat: Specify min-width for table

I am trying to display block the swissbiopic to appear
bigger, but it is not working in browsers, i.e. user agent
stylesheet is being ignored..

* Chore: Removes unnecessary code

* Chore: Adds github icon for more intuitive link

* Chore: Adds vertical divider

* Chore: Makes more descriptive labels

* Fix: Cell color and table match

This is was a bit tricky to fix.
Previous implementation adds colors to a list of already
exisiting color and subcellular location mapping.
This could get heavy when a user starts adding and
changing a lot of subcellular locations and colors.
I instead updated the list with the new colors,
adding only new colors whena  new subcellular location
is added too.

* Fix: Removes non-existent object

* Chore: Uses correct subcellular spelling

* Chore: Removes non-working css style

* Feat: Makes inputs compact segments

This reduces their sizes and the size they occupy.

* Chore: Rearranges color input, table, and cell

I think this arrangment provides the most space for the
cell. It would still be best if we can force #swissbiopic
to display: block.

* Chore: Suggests missing packages for CI

* Chore: Fixes spelling and updates WORDLIST

* Chore: Styles a lot of lines  to pass lintr

* Feat: Adds test via shinytest2

* Chore: Remove duplicate biocViews field

* Feat: Creates test for app using shinytest2

* Chore: Adds Recle as author; updates NEWS

* Chore: Update README.Rmd

I can't update README.md because knitr throws an error
when I try to render.
It seems to be caused by webshot.

* Feat: Updates unit tests for drawCell

The main test should be for drawCellShiny.
These unit tests are just to make sure the inputs are
properly sent to the htmlwidget.

* Chore: Removes unused helper function

This was actually causing lint errors since uniprot is not
defined globally.
`find_sc` function is not used anywhere, so it should be
safe to remove.

* Chore: Fixes lint settings and removes lint errors

* Chore: Adds shinytest2 as suggested dependency

This is for the app tests.

* Chore: Updates README and vignette

* Chore: Removes unneeded test

`edit_jscode` should no longer be used in drawCell.

* Chore: Adds colourpicker as package dependency

Otherwise, drawCellShiny will not work and the tests
will also fail.

* Chore: Ignores files for R CMD CHECK

* Chore: Adds DT as drawCellShiny dependency

* Chore: Remove snaps from drawCell test

shinytest2 will rebuild and test anyway. I am not sure
if this is standard practice, but I think we can do this
for now.

* Chore: Ignores hidden file in inst

This will hopefully make rcmdcheck pass.

* Chore: Removes ::: call to pass R CMD CHECK

* Chore: Updates package dependencies in NAMESPACE

* Chore: Removes vignettes from Rbuildignore

* Chore: Removes explicit imports in NAMESPACE

This will produce a note in R CMD CHECK,
but it should pass.

* merged main

* feat: Added the functionality to copy the code to the clipboard

* feat: I also updated the README, the options were the ones causing the trouble.

* Chore: Rearranges download buttons

* Chore: Fixes lint errors

* Chore: Tries svglite

Not working.

* Feat: Tries shinyscreenshot

Unfortunately, this does not work.
Currently, it seems like shinyscreenshot is not yet able
to take a screenshot and download svg images.
I tried taking screenshot of other elements and the whole
app, and I can confirm only the svg image cannot be
captured.

* Feat: Uses switch for animal cells

I did this because I will use the same logic in the
clear colors button.

* Feat: Gives cell a min-width

This is to make sure that the cell stays at a relatively
viewable size because it has a tendency to shrink until
it cannot be seen.
I also added a toast when user clicks the button to
download the code to inform them that the code is
workin already in the clipboard.
Reverted the download logic as I can't make shinyscreenshot
work with svg.

* fix: function moved to helpers.R

* Fix: Ignores testthat warnings

R CMD CHECK ignores the warnings and the warnings
are just shintest2 trying to compare snapshots,
so it should be safe to ignore.

* Fix: Do not take shiny screenshots

The json snapshots are good enough.
The png snapshots are not necessary and are inaccurate.
They also cause non-portable errors in R CMD CHECK.

* Chore: Updates WORDLIST to pass spellcheck

* Chore: Fixes style to pass lint

* Fix: Calls correct function

* Fix: Removes duplicate glue in imports

Ignores README_Files during build, otherwise it will
cause an error in R CMD CHECK.

* Fix: Removes remnants of merge

* Chore: Ensures README.Rmd renders

* merge master

* fix: linted package

* fix: Removed old unused function, which caused lint errors.

Co-authored-by: svalvaro <sanchezvillalba.alvaro@gmail.com>
Co-authored-by: Alvaro Sanchez <71273913+svalvaro@users.noreply.github.com>