Feat: Updates colourpicker label on cell click
This is to guide the user which subcellular location they
will color. At start up, the label of the colourpicker
will instruct them to click a subcellular location to
color.
After clicking or choosing a subcellular location,
the colourpicker label should change to instruct
the users to add color to the current selected
cell.